agent/requirement.txt
livekit
openai
mem0ai
python-dotenv
requests
